feat: Move GUI code to separate module and import it into main-dev.py
